split out barcode utils
